refactor(domain): split large service helpers
