+++ libs/incentive-system-domain/src/domain/domain-services/incentive.domain-service.ts
@@ -1,36 +1,38 @@
 import type {
-  PaymentMethod} from '../aggregates/incentive.aggregate.js';
-import {
-  Incentive,
   IncentiveStatus,
-  TriggerType,
-  Currency,
+  PaymentMethod,
 } from '../aggregates/incentive.aggregate.js';
-import { ContactInfo } from '../value-objects/index.js';
-import { IncentiveRules } from './incentive.rules.js';
+import type { ContactInfo } from '../value-objects/index.js';
 import type {
-  BatchPaymentItem,
-  IPIncentiveStatistics,
-  SystemIncentiveStatistics,
-  IIncentiveRepository,
-  IDomainEventBus,
   IAuditLogger,
-  IPaymentGateway} from '../../application/dtos/incentive.dto.js';
-import {
-  IncentiveCreationResult,
-  IncentiveValidationResult,
-  IncentiveApprovalResult,
-  IncentiveRejectionResult,
-  PaymentProcessingResult,
+  IDomainEventBus,
+  IIncentiveRepository,
+  IPaymentGateway,
   BatchPaymentResult,
+  IncentiveApprovalResult,
+  IncentiveCreationResult,
+  IncentiveRejectionResult,
   IncentiveStatsResult,
-  PendingIncentivesResult
+  IncentiveValidationResult,
+  PaymentProcessingResult,
+  PendingIncentivesResult,
 } from '../../application/dtos/incentive.dto.js';
+import { IncentiveCreationService } from './incentive/incentive-creation.service.js';
+import { IncentivePaymentService } from './incentive/incentive-payment.service.js';
+import { IncentiveReviewService } from './incentive/incentive-review.service.js';
+import { IncentiveStatisticsService } from './incentive/incentive-statistics.service.js';
+import { IncentiveValidationService } from './incentive/incentive-validation.service.js';
 
 /**
  * Provides incentive domain functionality.
  */
 export class IncentiveDomainService {
+  private readonly creationService: IncentiveCreationService;
+  private readonly validationService: IncentiveValidationService;
+  private readonly reviewService: IncentiveReviewService;
+  private readonly paymentService: IncentivePaymentService;
+  private readonly statisticsService: IncentiveStatisticsService;
+
   /**
    * Initializes a new instance of the Incentive Domain Service.
    * @param repository - The repository.
@@ -44,8 +46,32 @@ export class IncentiveDomainService {
     private readonly auditLogger: IAuditLogger,
     private readonly paymentGateway: IPaymentGateway,
   ) {
-  // Intentionally empty
-}
+    this.creationService = new IncentiveCreationService(
+      this.repository,
+      this.eventBus,
+      this.auditLogger,
+    );
+    this.validationService = new IncentiveValidationService(
+      this.repository,
+      this.eventBus,
+      this.auditLogger,
+    );
+    this.reviewService = new IncentiveReviewService(
+      this.repository,
+      this.eventBus,
+      this.auditLogger,
+    );
+    this.paymentService = new IncentivePaymentService(
+      this.repository,
+      this.eventBus,
+      this.auditLogger,
+      this.paymentGateway,
+    );
+    this.statisticsService = new IncentiveStatisticsService(
+      this.repository,
+      this.auditLogger,
+    );
+  }
 
   /**
    * 创建问卷完成激励
@@ -56,71 +82,12 @@ export class IncentiveDomainService {
     qualityScore: number,
     contactInfo: ContactInfo,
   ): Promise<IncentiveCreationResult> {
-    try {
-      // 获取IP今日激励数量
-      const todayIncentives = await this.repository.countTodayIncentives(ip);
-
-      // 验证创建资格
-      const eligibility = IncentiveRules.canCreateIncentive(
-        ip,
-        TriggerType.QUESTIONNAIRE_COMPLETION,
-        { questionnaireId, qualityScore },
-        todayIncentives,
-      );
-
-      if (!eligibility.isEligible) {
-        await this.auditLogger.logBusinessEvent('INCENTIVE_CREATION_DENIED', {
-          ip,
-          questionnaireId,
-          qualityScore,
-          errors: eligibility.errors,
-        });
-        return IncentiveCreationResult.failed(eligibility.errors);
-      }
-
-      // 创建激励
-      const incentive = Incentive.createQuestionnaireIncentive(
-        ip,
-        questionnaireId,
-        qualityScore,
-        contactInfo,
-      );
-
-      // 保存到存储
-      await this.repository.save(incentive);
-
-      // 发布领域事件
-      const events = incentive.getUncommittedEvents();
-      for(const event of events) {
-        await this.eventBus.publish(event);
-      }
-      incentive.markEventsAsCommitted();
-
-      // 记录审计日志
-      await this.auditLogger.logBusinessEvent('INCENTIVE_CREATED', {
-        incentiveId: incentive.getId().getValue(),
-        ip,
-        questionnaireId,
-        qualityScore,
-        rewardAmount: incentive.getRewardAmount(),
-        status: incentive.getStatus(),
-      });
-
-      return IncentiveCreationResult.success(incentive.getIncentiveSummary());
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('CREATE_QUESTIONNAIRE_INCENTIVE_ERROR', {
-        ip,
-        questionnaireId,
-        qualityScore,
-        error: errorMessage,
-      });
-      console.error('Error creating questionnaire incentive:', error);
-      return IncentiveCreationResult.failed([
-        'Internal error occurred while creating incentive',
-      ]);
-    }
+    return this.creationService.createQuestionnaireIncentive(
+      ip,
+      questionnaireId,
+      qualityScore,
+      contactInfo,
+    );
   }
 
   /**
@@ -131,73 +98,11 @@ export class IncentiveDomainService {
     referredIP: string,
     contactInfo: ContactInfo,
   ): Promise<IncentiveCreationResult> {
-    try {
-      // 验证推荐资格
-      const todayIncentives =
-        await this.repository.countTodayIncentives(referrerIP);
-      const eligibility = IncentiveRules.canCreateIncentive(
-        referrerIP,
-        TriggerType.REFERRAL,
-        { referredIP },
-        todayIncentives,
-      );
-
-      if (!eligibility.isEligible) {
-        await this.auditLogger.logBusinessEvent('REFERRAL_INCENTIVE_DENIED', {
-          referrerIP,
-          referredIP,
-          errors: eligibility.errors,
-        });
-        return IncentiveCreationResult.failed(eligibility.errors);
-      }
-
-      // 验证被推荐IP是否有效且未重复
-      const existingReferral = await this.repository.findReferralIncentive(
-        referrerIP,
-        referredIP,
-      );
-      if (existingReferral) {
-        return IncentiveCreationResult.failed([
-          'Referral incentive already exists for this IP pair',
-        ]);
-      }
-
-      // 创建推荐激励
-      const incentive = Incentive.createReferralIncentive(
-        referrerIP,
-        referredIP,
-        contactInfo,
-      );
-      await this.repository.save(incentive);
-
-      // 发布领域事件
-      const events = incentive.getUncommittedEvents();
-      for(const event of events) {
-        await this.eventBus.publish(event);
-      }
-      incentive.markEventsAsCommitted();
-
-      await this.auditLogger.logBusinessEvent('REFERRAL_INCENTIVE_CREATED', {
-        incentiveId: incentive.getId().getValue(),
-        referrerIP,
-        referredIP,
-        rewardAmount: incentive.getRewardAmount(),
-      });
-
-      return IncentiveCreationResult.success(incentive.getIncentiveSummary());
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('CREATE_REFERRAL_INCENTIVE_ERROR', {
-        referrerIP,
-        referredIP,
-        error: errorMessage,
-      });
-      console.error('Error creating referral incentive:', error);
-      return IncentiveCreationResult.failed([
-        'Internal error occurred while creating referral incentive',
-      ]);
-    }
+    return this.creationService.createReferralIncentive(
+      referrerIP,
+      referredIP,
+      contactInfo,
+    );
   }
 
   /**
@@ -206,49 +111,7 @@ export class IncentiveDomainService {
   public async validateIncentive(
     incentiveId: string,
   ): Promise<IncentiveValidationResult> {
-    try {
-      const incentive = await this.repository.findById(incentiveId);
-      if (!incentive) {
-        return IncentiveValidationResult.failed(['Incentive not found']);
-      }
-
-      // 执行验证
-      const validationResult = incentive.validateEligibility();
-
-      // 保存验证结果
-      await this.repository.save(incentive);
-
-      // 发布领域事件
-      const events = incentive.getUncommittedEvents();
-      for(const event of events) {
-        await this.eventBus.publish(event);
-      }
-      incentive.markEventsAsCommitted();
-
-      await this.auditLogger.logBusinessEvent('INCENTIVE_VALIDATED', {
-        incentiveId,
-        isValid: validationResult.isValid,
-        errors: validationResult.errors,
-      });
-
-      return IncentiveValidationResult.success({
-        incentiveId,
-        isValid: validationResult.isValid,
-        errors: validationResult.errors,
-        status: incentive.getStatus(),
-      });
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('VALIDATE_INCENTIVE_ERROR', {
-        incentiveId,
-        error: errorMessage,
-      });
-      console.error('Error validating incentive:', error);
-      return IncentiveValidationResult.failed([
-        'Internal error occurred while validating incentive',
-      ]);
-    }
+    return this.validationService.validateIncentive(incentiveId);
   }
 
   /**
@@ -258,47 +121,7 @@ export class IncentiveDomainService {
     incentiveId: string,
     reason: string,
   ): Promise<IncentiveApprovalResult> {
-    try {
-      const incentive = await this.repository.findById(incentiveId);
-      if (!incentive) {
-        return IncentiveApprovalResult.failed(['Incentive not found']);
-      }
-
-      // 执行批准
-      incentive.approveForProcessing(reason);
-      await this.repository.save(incentive);
-
-      // 发布领域事件
-      const events = incentive.getUncommittedEvents();
-      for(const event of events) {
-        await this.eventBus.publish(event);
-      }
-      incentive.markEventsAsCommitted();
-
-      await this.auditLogger.logBusinessEvent('INCENTIVE_APPROVED', {
-        incentiveId,
-        reason,
-        rewardAmount: incentive.getRewardAmount(),
-      });
-
-      return IncentiveApprovalResult.success({
-        incentiveId,
-        status: incentive.getStatus(),
-        rewardAmount: incentive.getRewardAmount(),
-      });
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('APPROVE_INCENTIVE_ERROR', {
-        incentiveId,
-        reason,
-        error: errorMessage,
-      });
-      console.error('Error approving incentive:', error);
-      return IncentiveApprovalResult.failed([
-        'Internal error occurred while approving incentive',
-      ]);
-    }
+    return this.reviewService.approveIncentive(incentiveId, reason);
   }
 
   /**
@@ -308,46 +131,7 @@ export class IncentiveDomainService {
     incentiveId: string,
     reason: string,
   ): Promise<IncentiveRejectionResult> {
-    try {
-      const incentive = await this.repository.findById(incentiveId);
-      if (!incentive) {
-        return IncentiveRejectionResult.failed(['Incentive not found']);
-      }
-
-      // 执行拒绝
-      incentive.reject(reason);
-      await this.repository.save(incentive);
-
-      // 发布领域事件
-      const events = incentive.getUncommittedEvents();
-      for(const event of events) {
-        await this.eventBus.publish(event);
-      }
-      incentive.markEventsAsCommitted();
-
-      await this.auditLogger.logBusinessEvent('INCENTIVE_REJECTED', {
-        incentiveId,
-        reason,
-      });
-
-      return IncentiveRejectionResult.success({
-        incentiveId,
-        status: incentive.getStatus(),
-        rejectionReason: reason,
-      });
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('REJECT_INCENTIVE_ERROR', {
-        incentiveId,
-        reason,
-        error: errorMessage,
-      });
-      console.error('Error rejecting incentive:', error);
-      return IncentiveRejectionResult.failed([
-        'Internal error occurred while rejecting incentive',
-      ]);
-    }
+    return this.reviewService.rejectIncentive(incentiveId, reason);
   }
 
   /**
@@ -358,103 +142,11 @@ export class IncentiveDomainService {
     paymentMethod: PaymentMethod,
     contactInfo?: ContactInfo,
   ): Promise<PaymentProcessingResult> {
-    try {
-      const incentive = await this.repository.findById(incentiveId);
-      if (!incentive) {
-        return PaymentProcessingResult.failed(['Incentive not found']);
-      }
-
-      // 验证支付资格
-      const eligibility = IncentiveRules.canPayIncentive(incentive);
-      if (!eligibility.isEligible) {
-        return PaymentProcessingResult.failed(eligibility.errors);
-      }
-
-      // 验证支付方式兼容性
-      const actualContactInfo =
-        contactInfo || this.extractContactInfoFromIncentive(incentive);
-      const methodValidation =
-        IncentiveRules.validatePaymentMethodCompatibility(
-          paymentMethod,
-          actualContactInfo,
-        );
-      if (!methodValidation.isValid) {
-        return PaymentProcessingResult.failed(methodValidation.errors);
-      }
-
-      // 通过支付网关处理支付
-      const paymentRequest = {
-        amount: incentive.getRewardAmount(),
-        currency: Currency.CNY,
-        paymentMethod,
-        recipientInfo: this.buildRecipientInfo(actualContactInfo),
-        reference: incentiveId,
-      };
-
-      const gatewayResult =
-        await this.paymentGateway.processPayment(paymentRequest);
-
-      // Check if payment gateway succeeded
-      if (!gatewayResult.success) {
-        return PaymentProcessingResult.failed([
-          gatewayResult.error ?? 'Payment gateway error',
-        ]);
-      }
-
-      // 执行激励支付
-      const paymentResult = incentive.executePayment(
-        paymentMethod,
-        gatewayResult.transactionId,
-      );
-
-      if (paymentResult.success) {
-        await this.repository.save(incentive);
-
-        // 发布领域事件
-        const events = incentive.getUncommittedEvents();
-        for(const event of events) {
-          await this.eventBus.publish(event);
-        }
-        incentive.markEventsAsCommitted();
-
-        await this.auditLogger.logBusinessEvent('INCENTIVE_PAID', {
-          incentiveId,
-          amount: paymentResult.amount,
-          currency: paymentResult.currency,
-          paymentMethod,
-          transactionId: gatewayResult.transactionId,
-        });
-
-        return PaymentProcessingResult.success({
-          incentiveId,
-          transactionId: gatewayResult.transactionId,
-          amount: paymentResult.amount ?? 0,
-          currency: paymentResult.currency ?? Currency.CNY,
-          paymentMethod,
-          status: incentive.getStatus(),
-        });
-      } else {
-        await this.auditLogger.logBusinessEvent('INCENTIVE_PAYMENT_FAILED', {
-          incentiveId,
-          error: paymentResult.error,
-          paymentMethod,
-        });
-
-        return PaymentProcessingResult.failed([paymentResult.error ?? 'Unknown payment error']);
-      }
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('PROCESS_PAYMENT_ERROR', {
-        incentiveId,
-        paymentMethod,
-        error: errorMessage,
-      });
-      console.error('Error processing payment:', error);
-      return PaymentProcessingResult.failed([
-        'Internal error occurred while processing payment',
-      ]);
-    }
+    return this.paymentService.processPayment(
+      incentiveId,
+      paymentMethod,
+      contactInfo,
+    );
   }
 
   /**
@@ -464,119 +156,7 @@ export class IncentiveDomainService {
     incentiveIds: string[],
     paymentMethod: PaymentMethod,
   ): Promise<BatchPaymentResult> {
-    try {
-      // 获取所有激励
-      const incentives = await this.repository.findByIds(incentiveIds);
-      if (incentives.length === 0) {
-        return BatchPaymentResult.failed(['No valid incentives found']);
-      }
-
-      // 验证批量支付
-      const batchValidation = IncentiveRules.validateBatchPayment(incentives);
-      if (!batchValidation.isValid) {
-        return BatchPaymentResult.failed(batchValidation.errors);
-      }
-
-      // 处理每个激励的支付
-      const results: BatchPaymentItem[] = [];
-      let successCount = 0;
-      let totalPaidAmount = 0;
-
-      for(const incentive of incentives) {
-        const eligibility = IncentiveRules.canPayIncentive(incentive);
-        if (!eligibility.isEligible) {
-          results.push({
-            incentiveId: incentive.getId().getValue(),
-            success: false,
-            error: eligibility.errors.join(', '),
-          });
-          continue;
-        }
-
-        try {
-          const contactInfo = this.extractContactInfoFromIncentive(incentive);
-          const paymentRequest = {
-            amount: incentive.getRewardAmount(),
-            currency: Currency.CNY,
-            paymentMethod,
-            recipientInfo: this.buildRecipientInfo(contactInfo),
-            reference: incentive.getId().getValue(),
-          };
-
-          const gatewayResult =
-            await this.paymentGateway.processPayment(paymentRequest);
-          const paymentResult = incentive.executePayment(
-            paymentMethod,
-            gatewayResult.transactionId,
-          );
-
-          if (paymentResult.success) {
-            await this.repository.save(incentive);
-
-            // 发布领域事件
-            const events = incentive.getUncommittedEvents();
-            for(const event of events) {
-              await this.eventBus.publish(event);
-            }
-            incentive.markEventsAsCommitted();
-
-            results.push({
-              incentiveId: incentive.getId().getValue(),
-              success: true,
-              transactionId: gatewayResult.transactionId,
-              amount: paymentResult.amount ?? 0,
-            });
-
-            successCount++;
-            totalPaidAmount += paymentResult.amount ?? 0;
-          } else {
-            results.push({
-              incentiveId: incentive.getId().getValue(),
-              success: false,
-              error: paymentResult.error ?? 'Unknown error',
-            });
-          }
-        } catch (paymentError) {
-          const errorMessage =
-            paymentError instanceof Error
-              ? paymentError.message
-              : 'Payment error';
-          results.push({
-            incentiveId: incentive.getId().getValue(),
-            success: false,
-            error: errorMessage,
-          });
-        }
-      }
-
-      await this.auditLogger.logBusinessEvent('BATCH_PAYMENT_PROCESSED', {
-        totalIncentives: incentiveIds.length,
-        successCount,
-        failureCount: incentiveIds.length - successCount,
-        totalPaidAmount,
-        paymentMethod,
-      });
-
-      return BatchPaymentResult.success({
-        totalIncentives: incentiveIds.length,
-        successCount,
-        failureCount: incentiveIds.length - successCount,
-        totalPaidAmount,
-        results,
-      });
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('PROCESS_BATCH_PAYMENT_ERROR', {
-        incentiveIds,
-        paymentMethod,
-        error: errorMessage,
-      });
-      console.error('Error processing batch payment:', error);
-      return BatchPaymentResult.failed([
-        'Internal error occurred while processing batch payment',
-      ]);
-    }
+    return this.paymentService.processBatchPayment(incentiveIds, paymentMethod);
   }
 
   /**
@@ -586,35 +166,7 @@ export class IncentiveDomainService {
     ip?: string,
     timeRange?: { startDate: Date; endDate: Date },
   ): Promise<IncentiveStatsResult> {
-    try {
-      if (ip) {
-        // 获取特定IP的统计
-        if (!this.isValidIPAddress(ip)) {
-          return IncentiveStatsResult.failed(['Invalid IP address format']);
-        }
-
-        const incentives = await this.repository.findByIP(ip, timeRange);
-        const stats = this.calculateIPStatistics(ip, incentives);
-
-        return IncentiveStatsResult.success({ individual: stats });
-      } else {
-        // 获取系统整体统计
-        const systemStats = await this.calculateSystemStatistics(timeRange);
-        return IncentiveStatsResult.success({ system: systemStats });
-      }
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('GET_INCENTIVE_STATISTICS_ERROR', {
-        ip,
-        timeRange,
-        error: errorMessage,
-      });
-      console.error('Error getting incentive statistics:', error);
-      return IncentiveStatsResult.failed([
-        'Internal error occurred while getting statistics',
-      ]);
-    }
+    return this.statisticsService.getIncentiveStatistics(ip, timeRange);
   }
 
   /**
@@ -624,184 +176,6 @@ export class IncentiveDomainService {
     status?: IncentiveStatus,
     limit = 50,
   ): Promise<PendingIncentivesResult> {
-    try {
-      const incentives = await this.repository.findPendingIncentives(
-        status,
-        limit,
-      );
-      const prioritizedIncentives = incentives
-        .map((incentive) => ({
-          incentive: incentive.getIncentiveSummary(),
-          priority: IncentiveRules.calculateProcessingPriority(incentive),
-        }))
-        .sort((a, b) => b.priority.score - a.priority.score);
-
-      return PendingIncentivesResult.success(prioritizedIncentives);
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : 'Unknown error';
-      await this.auditLogger.logError('GET_PENDING_INCENTIVES_ERROR', {
-        status,
-        limit,
-        error: errorMessage,
-      });
-      console.error('Error getting pending incentives:', error);
-      return PendingIncentivesResult.failed([
-        'Internal error occurred while getting pending incentives',
-      ]);
-    }
-  }
-
-  private buildRecipientInfo(contactInfo: ContactInfo): {
-    name: string;
-    email: string | undefined;
-    phone: string | undefined;
-    address: undefined;
-    accountNumber: string | undefined;
-    bankCode: undefined;
-    metadata: { wechat: string | undefined; alipay: string | undefined };
-  } {
-    return {
-      name: contactInfo.email ?? contactInfo.phone ?? 'Incentive Recipient',
-      email: contactInfo.email,
-      phone: contactInfo.phone,
-      address: undefined,
-      accountNumber: contactInfo.alipay ?? contactInfo.wechat,
-      bankCode: undefined,
-      metadata: {
-        wechat: contactInfo.wechat,
-        alipay: contactInfo.alipay,
-      },
-    };
-  }
-
-  // 私有辅助方法
-  private extractContactInfoFromIncentive(_incentive: Incentive): ContactInfo {
-    // 从激励中提取联系信息的逻辑
-    // 在实际实现中，这应该从激励的接收者中获取联系信息
-    // 这里提供一个基本的实现来满足测试需要
-    return new ContactInfo({
-      email: 'test@example.com',
-      wechat: 'test_wechat',
-      alipay: 'test_alipay',
-    });
-  }
-
-  private calculateIPStatistics(
-    ip: string,
-    incentives: Incentive[],
-  ): IPIncentiveStatistics {
-    let totalAmount = 0;
-    let paidAmount = 0;
-    let pendingAmount = 0;
-
-    const statusCount = {
-      pending_validation: 0,
-      approved: 0,
-      paid: 0,
-      rejected: 0,
-    };
-
-    for(const incentive of incentives) {
-      const amount = incentive.getRewardAmount();
-      totalAmount += amount;
-
-      switch (incentive.getStatus()) {
-        case IncentiveStatus.PENDING_VALIDATION:
-          statusCount.pending_validation++;
-          pendingAmount += amount;
-          break;
-        case IncentiveStatus.APPROVED:
-          statusCount.approved++;
-          pendingAmount += amount;
-          break;
-        case IncentiveStatus.PAID:
-          statusCount.paid++;
-          paidAmount += amount;
-          break;
-        case IncentiveStatus.REJECTED:
-          statusCount.rejected++;
-          break;
-      }
-    }
-
-    return {
-      ip,
-      totalIncentives: incentives.length,
-      totalAmount,
-      paidAmount,
-      pendingAmount,
-      statusBreakdown: statusCount,
-      averageReward:
-        incentives.length > 0 ? totalAmount / incentives.length : 0,
-      lastIncentiveDate:
-        incentives.length > 0
-          ? Math.max(...incentives.map((i) => i.getCreatedAt().getTime()))
-          : undefined,
-    };
-  }
-
-  private async calculateSystemStatistics(timeRange?: {
-    startDate: Date;
-    endDate: Date;
-  }): Promise<SystemIncentiveStatistics> {
-    const allIncentives = await this.repository.findAll(timeRange);
-
-    let totalAmount = 0;
-    let paidAmount = 0;
-    const uniqueIPs = new Set<string>();
-
-    const statusCount = {
-      pending_validation: 0,
-      approved: 0,
-      paid: 0,
-      rejected: 0,
-    };
-
-    for(const incentive of allIncentives) {
-      totalAmount += incentive.getRewardAmount();
-      uniqueIPs.add(incentive.getRecipientIP());
-
-      switch (incentive.getStatus()) {
-        case IncentiveStatus.PENDING_VALIDATION:
-          statusCount.pending_validation++;
-          break;
-        case IncentiveStatus.APPROVED:
-          statusCount.approved++;
-          break;
-        case IncentiveStatus.PAID:
-          statusCount.paid++;
-          paidAmount += incentive.getRewardAmount();
-          break;
-        case IncentiveStatus.REJECTED:
-          statusCount.rejected++;
-          break;
-      }
-    }
-
-    return {
-      totalIncentives: allIncentives.length,
-      uniqueRecipients: uniqueIPs.size,
-      totalAmount,
-      paidAmount,
-      pendingAmount: totalAmount - paidAmount,
-      statusBreakdown: statusCount,
-      averageRewardPerIncentive:
-        allIncentives.length > 0 ? totalAmount / allIncentives.length : 0,
-      averageRewardPerIP: uniqueIPs.size > 0 ? totalAmount / uniqueIPs.size : 0,
-      conversionRate:
-        allIncentives.length > 0
-          ? (statusCount.paid / allIncentives.length) * 100
-          : 0,
-    };
-  }
-
-  private isValidIPAddress(ip: string): boolean {
-    if (!ip || typeof ip !== 'string') return false;
-    const ipRegex =
-      /^(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/;
-    return ipRegex.test(ip);
+    return this.statisticsService.getPendingIncentives(status, limit);
   }
 }
-
-// Import rules and infrastructure interfaces - these need to be created separately
